refactor: ThemeStore의 IPC 로직을 Infrastructure 레이어로 분리 (#2)
